+++ api/webhook_test.go
@@ -1098,44 +1098,3 @@ func TestIncomingWebhooks(t *testing.T) {
 		t.Fatal("should have failed - webhooks turned off")
 	}
 }
-
-func TestCommandWebhooks(t *testing.T) {
-	th := Setup().InitSystemAdmin()
-	Client := th.SystemAdminClient
-	team := th.SystemAdminTeam
-	channel := th.CreateChannel(Client, team)
-	user := th.CreateUser(Client)
-	LinkUserToTeam(user, team)
-
-	cmd := &model.Command{
-		CreatorId: user.Id,
-		TeamId:    team.Id,
-		URL:       "http://nowhere.com",
-		Method:    model.COMMAND_METHOD_POST,
-		Trigger:   "delayed"}
-
-	cmd = Client.Must(Client.CreateCommand(cmd)).Data.(*model.Command)
-	args := &model.CommandArgs{
-		TeamId:    team.Id,
-		UserId:    user.Id,
-		ChannelId: channel.Id,
-	}
-	hook, err := app.CreateCommandWebhook(cmd.Id, args)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if _, err := Client.DoPost("/hooks/commands/123123123123", "{\"text\":\"this is a test\"}", "application/json"); err.StatusCode != http.StatusNotFound {
-		t.Fatal("expected not-found for non-existent hook")
-	}
-
-	for i := 0; i < 5; i++ {
-		if _, err := Client.DoPost("/hooks/commands/"+hook.Id, "{\"text\":\"this is a test\"}", "application/json"); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	if _, err := Client.DoPost("/hooks/commands/"+hook.Id, "{\"text\":\"this is a test\"}", "application/json"); err == nil || err.StatusCode != http.StatusBadRequest {
-		t.Fatal("expected error for sixth usage")
-	}
-}
