+++ model/command_webhook.go
@@ -3,6 +3,10 @@
 
 package model
 
+import (
+	"net/http"
+)
+
 type CommandWebhook struct {
 	Id        string
 	CreateAt  int64
@@ -30,31 +34,31 @@ func (o *CommandWebhook) PreSave() {
 
 func (o *CommandWebhook) IsValid() *AppError {
 	if len(o.Id) != 26 {
-		return NewLocAppError("CommandWebhook.IsValid", "model.command_hook.id.app_error", nil, "")
+		return NewAppError("CommandWebhook.IsValid", "model.command_hook.id.app_error", nil, "", http.StatusBadRequest)
 	}
 
 	if o.CreateAt == 0 {
-		return NewLocAppError("CommandWebhook.IsValid", "model.command_hook.create_at.app_error", nil, "id="+o.Id)
+		return NewAppError("CommandWebhook.IsValid", "model.command_hook.create_at.app_error", nil, "id="+o.Id, http.StatusBadRequest)
 	}
 
 	if len(o.CommandId) != 26 {
-		return NewLocAppError("CommandWebhook.IsValid", "model.command_hook.command_id.app_error", nil, "")
+		return NewAppError("CommandWebhook.IsValid", "model.command_hook.command_id.app_error", nil, "", http.StatusBadRequest)
 	}
 
 	if len(o.UserId) != 26 {
-		return NewLocAppError("CommandWebhook.IsValid", "model.command_hook.user_id.app_error", nil, "")
+		return NewAppError("CommandWebhook.IsValid", "model.command_hook.user_id.app_error", nil, "", http.StatusBadRequest)
 	}
 
 	if len(o.ChannelId) != 26 {
-		return NewLocAppError("CommandWebhook.IsValid", "model.command_hook.channel_id.app_error", nil, "")
+		return NewAppError("CommandWebhook.IsValid", "model.command_hook.channel_id.app_error", nil, "", http.StatusBadRequest)
 	}
 
 	if len(o.RootId) != 0 && len(o.RootId) != 26 {
-		return NewLocAppError("CommandWebhook.IsValid", "model.command_hook.root_id.app_error", nil, "")
+		return NewAppError("CommandWebhook.IsValid", "model.command_hook.root_id.app_error", nil, "", http.StatusBadRequest)
 	}
 
 	if len(o.ParentId) != 0 && len(o.ParentId) != 26 {
-		return NewLocAppError("CommandWebhook.IsValid", "model.command_hook.parent_id.app_error", nil, "")
+		return NewAppError("CommandWebhook.IsValid", "model.command_hook.parent_id.app_error", nil, "", http.StatusBadRequest)
 	}
 
 	return nil
